requirements.txt
apache-airflow-providers-snowflake>=4.0.0
pandas